.mcp/vision-bridge/requirements.txt
fastmcp<3
requests
pillow